[#2] Refactor : 메인페이지 cleanup
1.Main.tsx 는 1개의 컴포넌트만 호출 및 스토리지 용도로만 사용되고 있습니다.
ProductsList 에 filter,List 두가지 section 으로 분리하여 코드를 정리하였습니다.

2. product.tsx 파일의 버튼 클릭 이벤트 함수로 props data 를 반영하는 코드로 정리하였습니다.
