+++ src/components/main/Product.tsx
@@ -28,15 +28,15 @@ function Product(props: IProduct) {
 
 	const { isOpen, onOpen, onClose } = useDisclosure();
 
-	const handleReservation = (product: IProduct) => {
+	const handleReservation = () => {
 		let reservations = JSON.parse(localStorage.getItem('reservations') as string);
 
-		if (reservations.map((reservation: IReservation) => reservation.idx).includes(product.idx)) {
+		if (reservations.map((reservation: IReservation) => reservation.idx).includes(idx)) {
 			reservations = reservations.map((reservation: IReservation) =>
-				reservation.idx === product.idx ? { ...reservation, count: reservation.count + 1 } : reservation
+				reservation.idx === idx ? { ...reservation, count: reservation.count + 1 } : reservation
 			);
 		} else {
-			reservations = [...reservations, { ...product, count: 1 }];
+			reservations = [...reservations, { ...props, count: 1 }];
 		}
 		localStorage.setItem('reservations', JSON.stringify(reservations));
 
@@ -66,18 +66,7 @@ function Product(props: IProduct) {
 							_hover={{ bg: 'blue.300' }}
 							color="white"
 							boxShadow="2xl"
-							onClick={() =>
-								handleReservation({
-									idx,
-									name,
-									mainImage,
-									price,
-									spaceCategory,
-									description,
-									maximumPurchases,
-									registrationDate,
-								})
-							}
+							onClick={handleReservation}
 						>
 							예약하기
 						</Button>
